+++ Source/JNA/waffle-jna/src/main/java/waffle/windows/auth/impl/WindowsAuthProviderImpl.java
@@ -49,9 +49,13 @@ import waffle.windows.auth.IWindowsSecurityContext;
  * @author dblock[at]dblock[dot]org
  */
 public class WindowsAuthProviderImpl implements IWindowsAuthProvider {
-
+    private static class ContinueContext {
+        private CtxtHandle continueHandle;
+        private IWindowsCredentialsHandle serverCredential;
+    }
+    
     /** The continue contexts. */
-    private final Cache<String, CtxtHandle> continueContexts;
+    private final Cache<String, ContinueContext> continueContexts;
 
     /**
      * Instantiates a new windows auth provider impl.
@@ -84,27 +88,32 @@ public class WindowsAuthProviderImpl implements IWindowsAuthProvider {
             throw new Win32Exception(WinError.SEC_E_INVALID_TOKEN);
         }
 
-        final IWindowsCredentialsHandle serverCredential = new WindowsCredentialsHandleImpl(null,
-                Sspi.SECPKG_CRED_INBOUND, securityPackage);
-        serverCredential.initialize();
+        CtxtHandle continueHandle = null;
+        IWindowsCredentialsHandle serverCredential = null;
+        ContinueContext continueContext = continueContexts.asMap().get(connectionId);
+        if(continueContext != null) {
+            continueHandle = continueContext.continueHandle;
+            serverCredential = continueContext.serverCredential;
+        }
+        else {
+            serverCredential = new WindowsCredentialsHandleImpl(null, Sspi.SECPKG_CRED_INBOUND, securityPackage);
+            serverCredential.initialize();
+        }
 
         WindowsSecurityContextImpl sc;
 
         int rc;
         int tokenSize = Sspi.MAX_TOKEN_SIZE;
 
-        CtxtHandle continueContext;
-        SecBufferDesc pbServerToken;
-        SecBufferDesc pbClientToken;
-        final IntByReference pfClientContextAttr = new IntByReference();
-        final CtxtHandle phNewServerContext = new CtxtHandle();
         do {
-            pbServerToken = new SecBufferDesc(Sspi.SECBUFFER_TOKEN, tokenSize);
-            pbClientToken = new SecBufferDesc(Sspi.SECBUFFER_TOKEN, token);
+        	SecBufferDesc pbServerToken = new SecBufferDesc(Sspi.SECBUFFER_TOKEN, tokenSize);
+        	SecBufferDesc pbClientToken = new SecBufferDesc(Sspi.SECBUFFER_TOKEN, token);
+        	IntByReference pfClientContextAttr = new IntByReference();
 
             continueContext = this.continueContexts.asMap().get(connectionId);
 
-            rc = Secur32.INSTANCE.AcceptSecurityContext(serverCredential.getHandle(), continueContext, pbClientToken,
+            CtxtHandle phNewServerContext = new CtxtHandle();
+            rc = Secur32.INSTANCE.AcceptSecurityContext(serverCredential.getHandle(), continueHandle, pbClientToken,
                     Sspi.ISC_REQ_CONNECTION, Sspi.SECURITY_NATIVE_DREP, phNewServerContext, pbServerToken,
                     pfClientContextAttr, null);
 
@@ -117,7 +126,7 @@ public class WindowsAuthProviderImpl implements IWindowsAuthProvider {
                 case WinError.SEC_E_BUFFER_TOO_SMALL:
                     tokenSize += Sspi.MAX_TOKEN_SIZE;
                     sc.dispose();
-                    WindowsSecurityContextImpl.dispose(continueContext);
+                    WindowsSecurityContextImpl.dispose(continueHandle);
                     break;
                 case WinError.SEC_E_OK:
                     // the security context received from the client was accepted
@@ -131,13 +140,16 @@ public class WindowsAuthProviderImpl implements IWindowsAuthProvider {
                     break;
                 case WinError.SEC_I_CONTINUE_NEEDED:
                     // the server must send the output token to the client and wait for a returned token
-                    this.continueContexts.put(connectionId, phNewServerContext);
+                    continueContext = new ContinueContext();
+                    continueContext.serverCredential = serverCredential;
+                    continueContext.continueHandle = phNewServerContext;
+                    this.continueContexts.put(connectionId, continueContext);
                     sc.setToken(pbServerToken.getBytes() == null ? new byte[0] : pbServerToken.getBytes().clone());
                     sc.setContinue(true);
                     break;
                 default:
                     sc.dispose();
-                    WindowsSecurityContextImpl.dispose(continueContext);
+                    WindowsSecurityContextImpl.dispose(continueHandle);
                     this.continueContexts.asMap().remove(connectionId);
                     throw new Win32Exception(rc);
             }