+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Rework .net build to be mostly automatic using nuget
 * Change .net target to more modern .net 4.0 framework
 * [#296](https://github.com/dblock/waffle/pull/296): Added Tomcat 9 support.
+* [#268](https://github.com/dblock/waffle/pull/301): Cannot log in automatically on machine where Tomcat service is running
 
 1.8.0 (09/10/15)
 ================